rename polynomial folder
