+++ numba/np/polynomial/polynomial_core.py
