+++ numba/np/polynomial/polynomial_methods.py
